Re-Starting the Project
I've realised that my authentication is not properly working due to me not selecting the appropriate options when first creating this .NET Blazor project.
The employee and task management options are still included, but I'm starting the whole user authentication once again. Hopefully, this time it will work with no problems!